+++ lib/categories/providers/categories_provider.dart
@@ -1,8 +1,8 @@
-import 'package:expense_categoriser/categories/utils/models/category.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:isar/isar.dart';
 import 'package:path_provider/path_provider.dart';
 
+import '../../utils/models/categories/category.dart';
 import '../../reports/ui/uncategorised_item_row.dart';
 
 class CategoriesNotifier extends StateNotifier<List<Category>> {
