+++ lib/reports/ui/uncategorised_item_row.dart
@@ -1,8 +1,8 @@
-import 'package:expense_categoriser/categories/utils/models/category.dart';
 import 'package:expense_categoriser/reports/ui/selectable_words.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../utils/models/categories/category.dart';
 import '../utils/models/transaction.dart';
 import '../utils/models/selectable_word_item.dart';
