+++ lib/reports/view_models/report_view_model.dart
@@ -1,9 +1,9 @@
 import 'package:expense_categoriser/categories/services/categoriser.dart';
-import 'package:expense_categoriser/categories/utils/models/category.dart';
 import 'package:expense_categoriser/csv_files/services/csv_reader_service.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../utils/models/categories/category.dart';
 import '../utils/models/transaction.dart';
 import '../utils/models/report.dart';
